requirements.txt
django==3.1.6
djangorestframework==3.12.2
Pillow==5.3.0
pytz==2018.5
pip~=21.0
setuptools~=52.0.0
knox==0.1.11